Moved blueprint creation scripts to their own file
